Move NameContainsKeywordsPredicate to item package
